Move the scene component inside the FluidRenderer class
