+++ packages/dev/core/src/Rendering/fluidRenderer/fluidRenderer.ts
@@ -1,9 +1,13 @@
-import type { Scene } from "core/scene";
+import { Scene } from "core/scene";
 import type { Engine } from "core/Engines/engine";
 import type { FloatArray, Nullable } from "core/types";
 import type { Observer } from "core/Misc/observable";
 import type { Camera } from "core/Cameras/camera";
 import type { IParticleSystem } from "core/Particles/IParticleSystem";
+import type { ISceneComponent } from "core/sceneComponent";
+import { SceneComponentConstants } from "core/sceneComponent";
+import type { SmartArrayNoDuplicate } from "core/Misc/smartArray";
+import type { RenderTargetTexture } from "core/Materials/Textures/renderTargetTexture";
 
 import type { FluidRenderingObject } from "./fluidRenderingObject";
 import { FluidRenderingObjectParticleSystem } from "./fluidRenderingObjectParticleSystem";
@@ -20,6 +24,119 @@ import "../../Shaders/fluidRenderingParticleDiffuse.fragment";
 import "../../Shaders/fluidRenderingBilateralBlur.fragment";
 import "../../Shaders/fluidRenderingStandardBlur.fragment";
 import "../../Shaders/fluidRenderingRender.fragment";
+
+declare module "core/abstractScene" {
+    export interface AbstractScene {
+        /** @internal (Backing field) */
+        _fluidRenderer: Nullable<FluidRenderer>;
+
+        /**
+         * Gets or Sets the fluid renderer associated to the scene.
+         */
+        fluidRenderer: Nullable<FluidRenderer>;
+
+        /**
+         * Enables the fluid renderer and associates it with the scene
+         * @returns the FluidRenderer
+         */
+        enableFluidRenderer(): Nullable<FluidRenderer>;
+
+        /**
+         * Disables the fluid renderer associated with the scene
+         */
+        disableFluidRenderer(): void;
+    }
+}
+
+Object.defineProperty(Scene.prototype, "fluidRenderer", {
+    get: function (this: Scene) {
+        return this._fluidRenderer;
+    },
+    set: function (this: Scene, value: Nullable<FluidRenderer>) {
+        this._fluidRenderer = value;
+    },
+    enumerable: true,
+    configurable: true,
+});
+
+Scene.prototype.enableFluidRenderer = function (): Nullable<FluidRenderer> {
+    if (this._fluidRenderer) {
+        return this._fluidRenderer;
+    }
+
+    this._fluidRenderer = new FluidRenderer(this);
+
+    return this._fluidRenderer;
+};
+
+Scene.prototype.disableFluidRenderer = function (): void {
+    this._fluidRenderer?.dispose();
+    this._fluidRenderer = null;
+};
+
+/**
+ * Defines the fluid renderer scene component responsible to render objects as fluids
+ */
+export class FluidRendererSceneComponent implements ISceneComponent {
+    /**
+     * The component name helpful to identify the component in the list of scene components.
+     */
+    public readonly name = SceneComponentConstants.NAME_FLUIDRENDERER;
+
+    /**
+     * The scene the component belongs to.
+     */
+    public scene: Scene;
+
+    /**
+     * Creates a new instance of the component for the given scene
+     * @param scene Defines the scene to register the component in
+     */
+    constructor(scene: Scene) {
+        this.scene = scene;
+    }
+
+    /**
+     * Registers the component in a given scene
+     */
+    public register(): void {
+        this.scene._gatherActiveCameraRenderTargetsStage.registerStep(
+            SceneComponentConstants.STEP_GATHERACTIVECAMERARENDERTARGETS_FLUIDRENDERER,
+            this,
+            this._gatherActiveCameraRenderTargets
+        );
+        this.scene._afterCameraDrawStage.registerStep(SceneComponentConstants.STEP_AFTERCAMERADRAW_FLUIDRENDERER, this, this._afterCameraDraw);
+    }
+
+    private _gatherActiveCameraRenderTargets(_renderTargets: SmartArrayNoDuplicate<RenderTargetTexture>): void {
+        this.scene.fluidRenderer?._prepareRendering();
+    }
+
+    private _afterCameraDraw(camera: Camera) {
+        this.scene.fluidRenderer?._render(camera);
+    }
+
+    /**
+     * Rebuilds the elements related to this component in case of
+     * context lost for instance.
+     */
+    public rebuild(): void {
+        if (this.scene._fluidRenderer) {
+            // Release resources first
+            this.scene.disableFluidRenderer();
+
+            // Re-enable
+            this.scene.enableFluidRenderer();
+        }
+    }
+
+    /**
+     * Disposes the component and the associated resources
+     */
+    public dispose(): void {
+        this.scene.disableFluidRenderer();
+    }
+}
 
 /**
  * An object rendered as a fluid.
@@ -38,9 +155,13 @@ export interface IFluidRenderingRenderObject {
  */
 export class FluidRenderer {
     /** @internal */
-    public static _SceneComponentInitialization: (scene: Scene) => void = (/*_*/) => {
-        throw `FluidRendererSceneComponent needs to be imported before as it contains a side-effect required by your code.`;
-    };
+    public static _SceneComponentInitialization(scene: Scene) {
+        let component = scene._getComponent(SceneComponentConstants.NAME_FLUIDRENDERER) as FluidRendererSceneComponent;
+        if (!component) {
+            component = new FluidRendererSceneComponent(scene);
+            scene._addComponent(component);
+        }
+    }
 
     private _scene: Scene;
     private _engine: Engine;
