+++ packages/dev/core/src/Rendering/fluidRenderer/index.ts
@@ -1,5 +1,4 @@
 export * from "./fluidRenderer";
-export * from "./fluidRendererSceneComponent";
 export * from "./fluidRenderingObject";
 export * from "./fluidRenderingObjectParticleSystem";
 export * from "./fluidRenderingObjectCustomParticles";
